+++ test/src/tsplines_tests.cpp
@@ -11,87 +11,87 @@ using namespace tsl;
 namespace tsl_tests {
 
 TEST(RotateTest, Zero) {
-    vec2 vec(0.0f, 0.0f);
+    vec2 vec(0, 0);
     for (uint8_t i = 0; i < 10; ++i) {
         EXPECT_EQ(vec, rotate(i, vec));
     }
 }
 
 TEST(RotateTest, Values1) {
-    vec2 vec(1.0f, 0.0f);
+    vec2 vec(1, 0);
     EXPECT_EQ(vec,               rotate(0, vec));
-    EXPECT_EQ(vec2(0.0f,  1.0f), rotate(1, vec));
-    EXPECT_EQ(vec2(-1.0f, 0.0f), rotate(2, vec));
-    EXPECT_EQ(vec2(0.0f, -1.0f), rotate(3, vec));
+    EXPECT_EQ(vec2(0,  1), rotate(1, vec));
+    EXPECT_EQ(vec2(-1, 0), rotate(2, vec));
+    EXPECT_EQ(vec2(0, -1), rotate(3, vec));
     EXPECT_EQ(vec,               rotate(4, vec));
 }
 
 TEST(RotateTest, Values2) {
-    vec2 vec(0.0f, 1.0f);
+    vec2 vec(0, 1);
     EXPECT_EQ(vec,               rotate(0, vec));
-    EXPECT_EQ(vec2(-1.0f, 0.0f), rotate(1, vec));
-    EXPECT_EQ(vec2(0.0f, -1.0f), rotate(2, vec));
-    EXPECT_EQ(vec2(1.0f,  0.0f), rotate(3, vec));
+    EXPECT_EQ(vec2(-1, 0), rotate(1, vec));
+    EXPECT_EQ(vec2(0, -1), rotate(2, vec));
+    EXPECT_EQ(vec2(1,  0), rotate(3, vec));
     EXPECT_EQ(vec,               rotate(4, vec));
 }
 
 TEST(RotateTest, Values3) {
-    vec2 vec(1.0f, 1.0f);
+    vec2 vec(1, 1);
     EXPECT_EQ(vec,                rotate(0, vec));
-    EXPECT_EQ(vec2(-1.0f,  1.0f), rotate(1, vec));
-    EXPECT_EQ(vec2(-1.0f, -1.0f), rotate(2, vec));
-    EXPECT_EQ(vec2(1.0f,  -1.0f), rotate(3, vec));
+    EXPECT_EQ(vec2(-1,  1), rotate(1, vec));
+    EXPECT_EQ(vec2(-1, -1), rotate(2, vec));
+    EXPECT_EQ(vec2(1,  -1), rotate(3, vec));
     EXPECT_EQ(vec,                rotate(4, vec));
 }
 
 TEST(RotateTest, Values4) {
-    vec2 vec(-1.0f, 0.0f);
+    vec2 vec(-1, 0);
     EXPECT_EQ(vec,               rotate(0, vec));
-    EXPECT_EQ(vec2(0.0f, -1.0f), rotate(1, vec));
-    EXPECT_EQ(vec2(1.0f,  0.0f), rotate(2, vec));
-    EXPECT_EQ(vec2(0.0f,  1.0f), rotate(3, vec));
+    EXPECT_EQ(vec2(0, -1), rotate(1, vec));
+    EXPECT_EQ(vec2(1,  0), rotate(2, vec));
+    EXPECT_EQ(vec2(0,  1), rotate(3, vec));
     EXPECT_EQ(vec,               rotate(4, vec));
 }
 
 TEST(RotateTest, Values5) {
-    vec2 vec(0.0f, -1.0f);
+    vec2 vec(0, -1);
     EXPECT_EQ(vec,               rotate(0, vec));
-    EXPECT_EQ(vec2(1.0f,  0.0f), rotate(1, vec));
-    EXPECT_EQ(vec2(0.0f,  1.0f), rotate(2, vec));
-    EXPECT_EQ(vec2(-1.0f, 0.0f), rotate(3, vec));
+    EXPECT_EQ(vec2(1,  0), rotate(1, vec));
+    EXPECT_EQ(vec2(0,  1), rotate(2, vec));
+    EXPECT_EQ(vec2(-1, 0), rotate(3, vec));
     EXPECT_EQ(vec,               rotate(4, vec));
 }
 
 TEST(RotateTest, Values6) {
-    vec2 vec(-1.0f, -1.0f);
+    vec2 vec(-1, -1);
     EXPECT_EQ(vec,               rotate(0, vec));
-    EXPECT_EQ(vec2(1.0f, -1.0f), rotate(1, vec));
-    EXPECT_EQ(vec2(1.0f,  1.0f), rotate(2, vec));
-    EXPECT_EQ(vec2(-1.0f, 1.0f), rotate(3, vec));
+    EXPECT_EQ(vec2(1, -1), rotate(1, vec));
+    EXPECT_EQ(vec2(1,  1), rotate(2, vec));
+    EXPECT_EQ(vec2(-1, 1), rotate(3, vec));
     EXPECT_EQ(vec,               rotate(4, vec));
 }
 
 TEST(RotateTest, Values7) {
-    vec2 vec(-1.0f, 1.0f);
+    vec2 vec(-1, 1);
     EXPECT_EQ(vec,                rotate(0, vec));
-    EXPECT_EQ(vec2(-1.0f, -1.0f), rotate(1, vec));
-    EXPECT_EQ(vec2(1.0f,  -1.0f), rotate(2, vec));
-    EXPECT_EQ(vec2(1.0f,   1.0f), rotate(3, vec));
+    EXPECT_EQ(vec2(-1, -1), rotate(1, vec));
+    EXPECT_EQ(vec2(1,  -1), rotate(2, vec));
+    EXPECT_EQ(vec2(1,   1), rotate(3, vec));
     EXPECT_EQ(vec,                rotate(4, vec));
 }
 
 TEST(RotateTest, Values8) {
-    vec2 vec(1.0f, -1.0f);
+    vec2 vec(1, -1);
     EXPECT_EQ(vec,                rotate(0, vec));
-    EXPECT_EQ(vec2(1.0f,   1.0f), rotate(1, vec));
-    EXPECT_EQ(vec2(-1.0f,  1.0f), rotate(2, vec));
-    EXPECT_EQ(vec2(-1.0f, -1.0f), rotate(3, vec));
+    EXPECT_EQ(vec2(1,   1), rotate(1, vec));
+    EXPECT_EQ(vec2(-1,  1), rotate(2, vec));
+    EXPECT_EQ(vec2(-1, -1), rotate(3, vec));
     EXPECT_EQ(vec,                rotate(4, vec));
 }
 
 class TransformTest : public ::testing::Test {
 protected:
-    TransformTest() : t1(1, 0, vec2(1, 1)), t2(0.5f, 1, vec2(0, 0)), t3(2, 3, vec2(4, -5)), v(2, 3) {}
+    TransformTest() : t1(1, 0, vec2(1, 1)), t2(0.5, 1, vec2(0, 0)), t3(2, 3, vec2(4, -5)), v(2, 3) {}
 
     // 1 * R^0 + (1, 1)
     transform t1;
@@ -108,7 +108,7 @@ protected:
 
 TEST_F(TransformTest, ApplyVector) {
     EXPECT_EQ(vec2(3, 4), t1.apply(v));
-    EXPECT_EQ(vec2(-1.5f, 1), t2.apply(v));
+    EXPECT_EQ(vec2(-1.5, 1), t2.apply(v));
     EXPECT_EQ(vec2(10, -9), t3.apply(v));
 }
 
@@ -151,14 +151,14 @@ protected:
         //         v6 ============ v7
 
         auto& hem = mesh.mesh;
-        vertex_handles.push_back(hem.add_vertex(vec3(0.0f, 0.0f, 0.0f)));
-        vertex_handles.push_back(hem.add_vertex(vec3(0.0f, 0.0f, 0.0f)));
-        vertex_handles.push_back(hem.add_vertex(vec3(0.0f, 0.0f, 0.0f)));
-        vertex_handles.push_back(hem.add_vertex(vec3(0.0f, 0.0f, 0.0f)));
-        vertex_handles.push_back(hem.add_vertex(vec3(0.0f, 0.0f, 0.0f)));
-        vertex_handles.push_back(hem.add_vertex(vec3(0.0f, 0.0f, 0.0f)));
-        vertex_handles.push_back(hem.add_vertex(vec3(0.0f, 0.0f, 0.0f)));
-        vertex_handles.push_back(hem.add_vertex(vec3(0.0f, 0.0f, 0.0f)));
+        vertex_handles.push_back(hem.add_vertex(vec3(0, 0, 0)));
+        vertex_handles.push_back(hem.add_vertex(vec3(0, 0, 0)));
+        vertex_handles.push_back(hem.add_vertex(vec3(0, 0, 0)));
+        vertex_handles.push_back(hem.add_vertex(vec3(0, 0, 0)));
+        vertex_handles.push_back(hem.add_vertex(vec3(0, 0, 0)));
+        vertex_handles.push_back(hem.add_vertex(vec3(0, 0, 0)));
+        vertex_handles.push_back(hem.add_vertex(vec3(0, 0, 0)));
+        vertex_handles.push_back(hem.add_vertex(vec3(0, 0, 0)));
 
         face_handles.push_back(hem.add_face({vertex_handles[1], vertex_handles[0], vertex_handles[3], vertex_handles[4]}));
         face_handles.push_back(hem.add_face({vertex_handles[2], vertex_handles[1], vertex_handles[4], vertex_handles[5]}));
@@ -181,7 +181,7 @@ TEST_F(TmeshWithTfaceTest, GetExtendedValence) {
 }
 
 TEST(GetSpanTest, DefaultKnotVector) {
-    array<float, 5> knot_vector = {-2, -1, 0, 1, 2};
+    array<double, 5> knot_vector = {-2, -1, 0, 1, 2};
 
     EXPECT_EQ(0, *tsplines::get_span(-2, knot_vector));
     EXPECT_EQ(0, *tsplines::get_span(-1.5, knot_vector));
